+++ gamevault/ViewModels/MainWindowViewModel.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Controls;
+using System.Windows.Documents;
 
 namespace gamevault.ViewModels
 {
@@ -30,6 +31,18 @@ namespace gamevault.ViewModels
             m_Installs = new InstallUserControl();
             m_Community = new CommunityUserControl();
             m_AdminConsole = new AdminConsoleUserControl();
+        }
+        public void UpdateTaskbarProgress()
+        {
+            List<int> downloadProgresses = new List<int>();
+            foreach (var download in DownloadsViewModel.Instance.DownloadedGames)
+            {
+                if (download.IsDownloading())
+                {
+                    downloadProgresses.Add(download.GetDownloadProgress());
+                }
+            }
+            TaskbarProgress = downloadProgresses.Count > 0 ? (downloadProgresses.Average() / 100) : 0.0;
         }
         #region Singleton
         private static MainWindowViewModel instance = null;
@@ -75,7 +88,8 @@ namespace gamevault.ViewModels
             }
         }
         #endregion
-        #region PrivateMembers        
+        #region PrivateMembers     
+        private double m_TaskbarProgress = 0;
         private int m_ActiveControlIndex = -1;
         private UserControl m_ActiveControl { get; set; }
         private LibraryUserControl m_Library { get; set; }
@@ -87,6 +101,11 @@ namespace gamevault.ViewModels
         private AdminConsoleUserControl m_AdminConsole { get; set; }
         #endregion
         public MainControl LastMainControl { get; set; }
+        public double TaskbarProgress
+        {
+            get { return m_TaskbarProgress; }
+            set { m_TaskbarProgress = value; OnPropertyChanged(); }
+        }
         public int ActiveControlIndex
         {
             get { return m_ActiveControlIndex; }